More bug fixes to hash map
One is a legitimate fix, the other is a temporary work-around around an
issue I don't have time to investigate now. It introduces a minor memory
leak, but the only time this leaking occurs is when loading
configuration that causes a setting in the hash to be overwritten. In
other words, this leak will not ever become significant unless you try
to load a config file with a million identical entries.