Fixed bug in save logic
The hashmap was deallocated at the end of session load, but I forgot to
recreate it when starting to write to file, resulting in Putty
attempting to use deallocated space. This is now fixed. There is still a
saving bug triggered by the cleanup logic in hashmap_add that I will
need to investigate.