Added support for colored bold fonts
Most terminals treat bold color and font as 2 independent options. Many
default to having both a a thicker font and different color for bold
fonts. Putty only lets you choose one or the other. This commit makes
color and font thickness independet of each other in the commit,
allowing the look that's more similar to other terminals. The gtkwin
portion of this was borrowed from a Debian thread, the windows portion
was implemented by me.