Fixed and defaulted to legacy containers
Hashmap still seems to cause some inconsistent behavior, given that I
planned this as a productivity tool first, chasing a rare seg fault is
not a productive use of my time. I will instead default to original
logic that's easier to debug. Also added a fix for a bug that seemed to
be in original storage logic (null key entry that isn't checked for).